Refuse a single-precision subprogram instead of emitting it in double
ELM's shr_spfn_mod has calerf_r4 beside calerf_r8, the same erf with
`integer, parameter :: rk = r4` and `real(rk)` throughout. The numpy
renderer emitted the r4 routine entirely in float64 -- `np.float64('0.5E0')`
for `0.5E0_rk`, `y = abs(x)` on doubles -- with nothing but `# SKIPPED
local` hoisting notes to show for it, while the constants fold refuses the
same fact in a single constant ("this fold has no single-precision
arithmetic"). A call to shr_spfn_erf_r4 would have returned a double erf
where the Fortran returns a single, and the bit-exact gate would have failed
on it with no note pointing at the kind.

Two causes, both fixed:

* the frontend typed `real(rk)` as UNKNOWN_REAL_KIND(rk) because the kind
  map was the module's only, and every consumer read an unknown kind as a
  double. `local_kind_aliases` now reads a routine's own integer kind
  parameters (`rk = r4`, `rk = selected_real_kind(12)`, `rk = 8` where a
  declaration spells it as a kind) against the module's map, the
  subprogram record carries them as `kind_map`, and the declarations, the
  result, the constants fold of the local parameters and the expression
  renderer's kind lookups all resolve through them. calerf_r8's
  declarations are float64 now, and its `0.5E0_rk` folds instead of being
  skipped as a kind the fold does not know.

* the prologue refuses a subprogram any of whose dummies, result, locals
  or local parameters is float32 or complex64 -- a named single kind, the
  bare default real, or an alias of either -- as P001 with every such
  entity named, and raises before a statement of the body can run. This is
  the existing single-precision policy applied at the routine's scale.

Four test fixtures that declared bare `real` dummies for rules unrelated to
precision now say `real(8)`. The numfor corpus baseline moves by the three
`real(sp)` routines of its strings unit (deferred 2 -> 5, already stopped
by static.rwset); no other case declares a single-precision entity.

Closes 97.


Carried-From: e898f10b641bf05ab53f088448f38e3b9699d309

+++ tests/test_numpy_translate.py
@@ -427,8 +427,8 @@ module search_mod
 contains
   subroutine first_above( n, z, zmax, k_found, k_scan )
     integer, intent(in) :: n
-    real, dimension(n), intent(in) :: z
-    real, intent(in) :: zmax
+    real(8), dimension(n), intent(in) :: z
+    real(8), intent(in) :: zmax
     integer, intent(out) :: k_found, k_scan
     integer :: k, kk
     do k = 1, n
@@ -491,8 +491,8 @@ module lapack_wrap
 contains
   subroutine band_solvex( n, a, x )
     integer, intent(in) :: n
-    real, intent(inout) :: a(n)
-    real, intent(out) :: x(n)
+    real(8), intent(inout) :: a(n)
+    real(8), intent(out) :: x(n)
     x = a
   end subroutine band_solvex
 end module lapack_wrap
@@ -505,9 +505,9 @@ module solver_mod
 contains
   subroutine solve( method, n, m, a, x )
     integer, intent(in) :: method, n, m
-    real, intent(inout) :: a(n)
-    real, intent(out) :: x(n)
-    real :: work(n, max(2, m))
+    real(8), intent(inout) :: a(n)
+    real(8), intent(out) :: x(n)
+    real(8) :: work(n, max(2, m))
     work = 0.0
     if ( method == 1 ) then
       call band_solvex( n, a, x )
@@ -565,17 +565,17 @@ module relay_mod
   implicit none
 contains
   subroutine inner( x, y, rc )
-    real, intent(in) :: x
-    real, intent(out) :: y
-    real, intent(out), optional :: rc
+    real(8), intent(in) :: x
+    real(8), intent(out) :: y
+    real(8), intent(out), optional :: rc
     y = 2.0 * x
     if ( present(rc) ) rc = 1.0 / x
   end subroutine inner
   subroutine outer( x, y, rc, scale )
-    real, intent(in) :: x
-    real, intent(out) :: y
-    real, intent(out), optional :: rc
-    real, intent(in), optional :: scale
+    real(8), intent(in) :: x
+    real(8), intent(out) :: y
+    real(8), intent(out), optional :: rc
+    real(8), intent(in), optional :: scale
     call inner( x, y, rc = rc )
     if ( present(scale) ) y = y * scale
   end subroutine outer
